move compatibility classes https://github.com/xibosignageltd/xibo-private/issues/210
